backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
google-genai
pydantic
requests
python-dotenv
deep-translator